requirements.txt
fastapi
uvicorn
python-multipart
pandas
openpyxl
supabase
openai
python-dotenv
pydantic
tiktoken